% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunctions.R
\docType{data}
\name{adduct_type_default}
\alias{adduct_type_default}
\title{AB', 'PA', 'user_defined'}
\format{
An object of class \code{numeric} of length 3.
}
\usage{
adduct_type_default
}
\description{
AB', 'PA', 'user_defined'
}
\keyword{datasets}
