% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubFunctions.R
\name{MakeAllMonosCombos}
\alias{MakeAllMonosCombos}
\title{Generate All Possible Monosaccharide Composition Combinations}
\usage{
MakeAllMonosCombos(
  monos_list,
  min_num_custom = NULL,
  max_num_custom = NULL,
  minmax_map
)
}
\arguments{
\item{monos_list}{A named numeric vector where names are monosaccharide symbols
(e.g., \code{Hex}, \code{HexNAc}) and values are their respective monoisotopic masses.}

\item{min_num_custom}{A named numeric vector specifying customized minimum counts
for specific monosaccharides. Overrides or supplements \code{minmax_map}. Default is \code{NULL}.}

\item{max_num_custom}{A named numeric vector specifying customized maximum counts
for specific monosaccharides. Overrides or supplements \code{minmax_map}. Default is \code{NULL}.}

\item{minmax_map}{A named list where each element is a numeric vector of length 2
(e.g., \code{list(Hex = c(3, 10))}), representing the default min and max range.}
}
\value{
A data frame where each row represents a unique glycan composition and
each column corresponds to a monosaccharide type from \code{monos_list}.
}
\description{
This function creates a comprehensive data frame of all potential monosaccharide
combinations based on defined minimum and maximum counts for each saccharide unit.
It is typically used to build a search space for glycan mass matching.
}
\details{
The function aligns the input ranges with the \code{monos_list}. It prioritizes
\code{custom} inputs over the \code{minmax_map}. If any monosaccharide in
\code{monos_list} lacks a defined range, the function will throw an error.
The final output is generated using \code{\link[base]{expand.grid}}, representing
every mathematical combination within the specified constraints.
}
\examples{
\dontrun{
monos_list <- c(Hex = 162.0528, HexNAc = 203.0794, Fuc = 146.0579)
minmax_map <- list(Hex = c(3, 5), HexNAc = c(2, 4), Fuc = c(0, 2))

# Generate combinations
combo_table <- MakeAllMonosCombos(
  monos_list = monos_list,
  minmax_map = minmax_map
)
head(combo_table)
}
}
